test: Initial container refactor
