+++ packages/testing/containers/n8n-start-stack.ts
@@ -3,15 +3,15 @@ import { parseArgs } from 'node:util';
 
 import { getDockerImageFromEnv } from './docker-image';
 import { DockerImageNotFoundError } from './docker-image-not-found-error';
-import type { N8NConfig, N8NStack } from './n8n-test-container-creation';
-import { createN8NStack } from './n8n-test-container-creation';
+import { BASE_PERFORMANCE_PLANS, isValidPerformancePlan } from './performance-plans';
 import {
 	KEYCLOAK_TEST_CLIENT_ID,
 	KEYCLOAK_TEST_CLIENT_SECRET,
 	KEYCLOAK_TEST_USER_EMAIL,
 	KEYCLOAK_TEST_USER_PASSWORD,
-} from './n8n-test-container-keycloak';
-import { BASE_PERFORMANCE_PLANS, isValidPerformancePlan } from './performance-plans';
+} from './services/keycloak';
+import type { N8NConfig, N8NStack } from './stack';
+import { createN8NStack } from './stack';
 
 // ANSI colors for terminal output
 const colors = {
@@ -44,7 +44,6 @@ ${colors.yellow}Usage:${colors.reset}
 ${colors.yellow}Options:${colors.reset}
   --postgres        Use PostgreSQL instead of SQLite
   --queue           Enable queue mode (requires PostgreSQL)
-  --no-task-runner  Disable external task runner (enabled by default)
   --source-control  Enable source control (Git) container for testing
   --oidc            Enable OIDC testing with Keycloak (requires PostgreSQL)
   --observability   Enable observability stack (VictoriaLogs + VictoriaMetrics + Vector)
@@ -77,9 +76,6 @@ ${colors.yellow}Examples:${colors.reset}
   ${colors.bright}# Queue mode (automatically uses PostgreSQL)${colors.reset}
   npm run stack --queue
 
-  ${colors.bright}# Without task runner (task runner is enabled by default)${colors.reset}
-  npm run stack --no-task-runner
-
   ${colors.bright}# With source control (Git) testing${colors.reset}
   npm run stack --postgres --source-control
 
@@ -109,7 +105,7 @@ ${Object.keys(BASE_PERFORMANCE_PLANS)
 
 ${colors.yellow}Notes:${colors.reset}
   • SQLite is the default database (no external dependencies)
-  • Task runner is enabled by default (mirrors production)
+  • Task runner is always enabled (mirrors production)
   • Queue mode requires PostgreSQL and enables horizontal scaling
   • Use --name for running multiple instances in parallel
   • Performance plans simulate cloud constraints (SQLite only, resource-limited)
@@ -124,7 +120,6 @@ async function main() {
 			help: { type: 'boolean', short: 'h' },
 			postgres: { type: 'boolean' },
 			queue: { type: 'boolean' },
-			'no-task-runner': { type: 'boolean' },
 			'source-control': { type: 'boolean' },
 			oidc: { type: 'boolean' },
 			observability: { type: 'boolean' },
@@ -145,10 +140,8 @@ async function main() {
 	}
 
 	// Build configuration
-	// Task runner is enabled by default; use --no-task-runner to disable
 	const config: N8NConfig = {
 		postgres: values.postgres ?? false,
-		taskRunner: values['no-task-runner'] ? false : undefined, // Default true, only set false if explicitly disabled
 		sourceControl: values['source-control'] ?? false,
 		oidc: values.oidc ?? false,
 		observability: values.observability ?? false,
@@ -156,7 +149,7 @@ async function main() {
 		projectName: values.name ?? `n8n-stack-${Math.random().toString(36).substring(7)}`,
 	};
 
-	// Handle queue mode
+	// Handle queue mode (mains > 1 or workers > 0)
 	if (values.queue ?? values.mains ?? values.workers) {
 		const mains = parseInt(values.mains ?? '1', 10);
 		const workers = parseInt(values.workers ?? '1', 10);
@@ -166,7 +159,8 @@ async function main() {
 			process.exit(1);
 		}
 
-		config.queueMode = { mains, workers };
+		config.mains = mains;
+		config.workers = workers;
 
 		if (!values.queue && (values.mains ?? values.workers)) {
 			log.warn('--mains and --workers imply queue mode');
@@ -192,7 +186,8 @@ async function main() {
 
 		config.resourceQuota = plan;
 		config.postgres = false; // Force SQLite for performance plans
-		config.queueMode = false; // Force single instance for performance plans
+		config.mains = 1; // Force single instance for performance plans
+		config.workers = 0;
 
 		log.info(
 			`Using ${planName} performance plan: ${plan.memory}GB RAM, ${plan.cpu} CPU cores (SQLite only)`,
@@ -255,20 +250,19 @@ async function main() {
 			console.log('');
 			log.header('Observability Stack (VictoriaObs)');
 			log.info(
-				`VictoriaLogs UI: ${colors.cyan}${stack.observability.victoriaLogs.queryEndpoint}/select/vmui${colors.reset}`,
+				`VictoriaLogs UI: ${colors.cyan}${stack.observability.meta.logs.queryEndpoint}/select/vmui${colors.reset}`,
 			);
 			log.info(
-				`VictoriaMetrics UI: ${colors.cyan}${stack.observability.victoriaMetrics.queryEndpoint}/vmui${colors.reset}`,
+				`VictoriaMetrics UI: ${colors.cyan}${stack.observability.meta.metrics.queryEndpoint}/vmui${colors.reset}`,
 			);
-			if (stack.observability.vector) {
-				log.success('Container logs collected by Vector (runs in background)');
-			}
+			// Vector is always started when observability is enabled in the new stack
+			log.success('Container logs collected by Vector (runs in background)');
 		}
 
 		if (stack.tracing) {
 			console.log('');
 			log.header('Tracing Stack (n8n-tracer + Jaeger)');
-			log.info(`Jaeger UI: ${colors.cyan}${stack.tracing.jaeger.uiUrl}${colors.reset}`);
+			log.info(`Jaeger UI: ${colors.cyan}${stack.tracing.meta.jaeger.uiUrl}${colors.reset}`);
 		}
 
 		console.log('');
@@ -285,27 +279,28 @@ function displayConfig(config: N8NConfig) {
 	const dockerImage = getDockerImageFromEnv();
 	log.info(`Docker image: ${dockerImage}`);
 
+	const mains = config.mains ?? 1;
+	const workers = config.workers ?? 0;
+	const isQueueMode = mains > 1 || workers > 0;
+
 	// Determine actual database
 	// eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
-	const usePostgres = config.postgres || config.queueMode || config.oidc;
+	const usePostgres = config.postgres || isQueueMode || config.oidc;
 	log.info(`Database: ${usePostgres ? 'PostgreSQL' : 'SQLite'}`);
 
-	if (config.queueMode) {
-		const qm = typeof config.queueMode === 'boolean' ? { mains: 1, workers: 1 } : config.queueMode;
-		log.info(`Queue mode: ${qm.mains} main(s), ${qm.workers} worker(s)`);
+	if (isQueueMode) {
+		log.info(`Queue mode: ${mains} main(s), ${workers} worker(s)`);
 		if (!config.postgres) {
 			log.info('(PostgreSQL automatically enabled for queue mode)');
 		}
-		if (qm.mains && qm.mains > 1) {
+		if (mains > 1) {
 			log.info('(load balancer will be configured)');
 		}
 	} else {
 		log.info('Queue mode: disabled');
 	}
 
-	// Display task runner status (enabled by default)
-	const taskRunnerEnabled = config.taskRunner ?? true;
-	log.info(`Task runner: ${taskRunnerEnabled ? 'enabled (default)' : 'disabled'}`);
+	log.info('Task runner: enabled');
 
 	// Display source control status
 	if (config.sourceControl) {
@@ -319,7 +314,7 @@ function displayConfig(config: N8NConfig) {
 	// Display OIDC status
 	if (config.oidc) {
 		log.info('OIDC: enabled (Keycloak)');
-		if (!config.postgres && !config.queueMode) {
+		if (!config.postgres && !isQueueMode) {
 			log.info('(PostgreSQL automatically enabled for OIDC)');
 		}
 	} else {
