+++ packages/testing/playwright/tests/e2e/settings/environments/source-control.spec.ts
@@ -1,5 +1,3 @@
-import { addGiteaBranch, addGiteaRepo } from 'n8n-containers/n8n-test-container-gitea';
-
 import { expect, test } from '../../../../fixtures/base';
 import type { n8nPage } from '../../../../pages/n8nPage';
 import {
@@ -23,15 +21,14 @@ test.describe('Source Control Settings @capability:source-control', () => {
 	let repoUrl: string;
 	let repoName: string;
 
-	test.beforeEach(async ({ n8n, n8nContainer }) => {
+	test.beforeEach(async ({ n8n, chaos }) => {
 		await n8n.api.enableFeature('sourceControl');
-		await initSourceControl({ n8n, n8nContainer });
+		const gitea = chaos.services.gitea;
+		await initSourceControl({ n8n, gitea });
 
 		// Create unique repo with branches via API (not UI)
 		repoName = generateUniqueRepoName();
-		const giteaContainer = n8nContainer.containers.find((c) => c.getName().includes('gitea'));
-
-		await addGiteaRepo(giteaContainer!, repoName, 'giteaadmin', 'giteapassword');
+		await gitea.createRepo(repoName);
 
 		repoUrl = buildRepoUrl(repoName);
 	});
@@ -55,11 +52,11 @@ test.describe('Source Control Settings @capability:source-control', () => {
 		await expect(n8n.sideBar.getSourceControlConnectedIndicator()).toBeVisible();
 	});
 
-	test('should switch between branches', async ({ n8n, n8nContainer }) => {
-		const giteaContainer = n8nContainer.containers.find((c) => c.getName().includes('gitea'));
-		await addGiteaBranch(giteaContainer!, repoName, 'development', 'giteaadmin', 'giteapassword');
-		await addGiteaBranch(giteaContainer!, repoName, 'staging', 'giteaadmin', 'giteapassword');
-		await addGiteaBranch(giteaContainer!, repoName, 'production', 'giteaadmin', 'giteapassword');
+	test('should switch between branches', async ({ n8n, chaos }) => {
+		const gitea = chaos.services.gitea;
+		await gitea.createBranch(repoName, 'development');
+		await gitea.createBranch(repoName, 'staging');
+		await gitea.createBranch(repoName, 'production');
 
 		await n8n.api.sourceControl.connect({ repositoryUrl: repoUrl });
