+++ packages/testing/playwright/tests/e2e/settings/log-streaming/log-streaming-observability.spec.ts
@@ -28,11 +28,11 @@ test.describe('Log Streaming to VictoriaLogs @capability:observability', () => {
 	}) => {
 		// Get observability stack from the container
 		const obsStack = n8nContainer.observability!;
-		const obs = new ObservabilityHelper(obsStack);
+		const obs = new ObservabilityHelper(obsStack.meta);
 
 		// Configure syslog destination pointing to VictoriaLogs
 		const destination = await api.createSyslogDestination({
-			...obsStack.victoriaLogs.syslog, // host, port
+			...obsStack.meta.logs.syslog, // host, port
 			...SYSLOG_DEFAULTS, // protocol, facility, app_name
 			label: 'VictoriaLogs Test Destination',
 		});
@@ -60,7 +60,7 @@ test.describe('Log Streaming to VictoriaLogs @capability:observability', () => {
 	test('should query metrics from VictoriaMetrics', async ({ api, n8nContainer }) => {
 		// Get observability stack from the container
 		const obsStack = n8nContainer.observability!;
-		const obs = new ObservabilityHelper(obsStack);
+		const obs = new ObservabilityHelper(obsStack.meta);
 
 		// Import and activate a webhook workflow to generate metrics
 		const { webhookPath, workflowId } = await api.workflows.importWorkflowFromFile(
