+++ packages/testing/playwright/tests/e2e/settings/log-streaming/log-streaming-ui-e2e.spec.ts
@@ -23,7 +23,7 @@ test.describe('Log Streaming UI E2E @capability:observability', () => {
 		n8nContainer,
 	}) => {
 		const obsStack = n8nContainer.observability!;
-		const obs = new ObservabilityHelper(obsStack);
+		const obs = new ObservabilityHelper(obsStack.meta);
 
 		// ========== STEP 1: Configure Log Streaming via UI ==========
 		await n8n.navigate.toLogStreaming();
@@ -32,8 +32,8 @@ test.describe('Log Streaming UI E2E @capability:observability', () => {
 		// Create syslog destination pointing to VictoriaLogs
 		await n8n.settingsLogStreaming.createSyslogDestination({
 			name: 'VictoriaLogs E2E Test',
-			host: obsStack.victoriaLogs.syslog.host,
-			port: obsStack.victoriaLogs.syslog.port,
+			host: obsStack.meta.logs.syslog.host,
+			port: obsStack.meta.logs.syslog.port,
 		});
 		// Send test event
 		await n8n.settingsLogStreaming.sendTestEvent();
